requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# AI and ML dependencies
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2

# Data processing
pandas==2.1.4
numpy==1.24.3

# Web framework
jinja2==3.1.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Configuration
pyyaml==6.0 